Combined the hop report page and the hop entry page. End goal is that entering information in the edit page will auto-filter the hop list and if none match you can enter a new one.
There is an error occuring that causes the grid not load if I launch the client project directly. If the server project is launched then the page loads but still displays an error. Need to track down what in the setup is causing this.

Fixed how the DBContext is added to DI. So far there is no parameter passed to the DBContext constructor so there is no need to add an options object.